Add DataSchemas for parameter types on naptime handlers
This change adds a new `typeSchema` field to the Parameter, which contains a DataSchema (serialized as a DataMap) that defines the parameter type. This allows the type of a parameter to be serialized in a better way than a string, so it can be parsed by the GraphQL assembler and used in schema generation. (Presently, the scala typename is being parsed with a regex, but that’s terrible)